Nueva administración "Departamento".

+++ src/main/java/es/mira/progesin/persistence/repositories/IPuestoTrabajoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import es.mira.progesin.persistence.entities.PuestoTrabajo;
 
-public interface IPuestoTrabajoRepository extends CrudRepository<PuestoTrabajo, Integer> {
+public interface IPuestoTrabajoRepository extends CrudRepository<PuestoTrabajo, Long> {
     /**
      * 
      * @return Puestos de trabajo sin fecha de baja, es decir activos
